Export VespaZookeeperServer from a separate package
