+++ zookeeper-server/zookeeper-server-common/src/main/java/com/yahoo/vespa/zookeeper/Reconfigurer.java
@@ -5,6 +5,7 @@ import com.yahoo.cloud.config.ZookeeperServerConfig;
 import com.yahoo.component.AbstractComponent;
 import com.yahoo.component.annotation.Inject;
 import com.yahoo.protect.Process;
+import com.yahoo.vespa.zookeeper.server.VespaZooKeeperServer;
 import com.yahoo.yolean.Exceptions;
 
 import java.time.Duration;
