+++ zookeeper-server/zookeeper-server-common/src/main/java/com/yahoo/vespa/zookeeper/ZooKeeperRunner.java
@@ -4,6 +4,7 @@ package com.yahoo.vespa.zookeeper;
 import com.yahoo.cloud.config.ZookeeperServerConfig;
 import com.yahoo.concurrent.DaemonThreadFactory;
 import com.yahoo.protect.Process;
+import com.yahoo.vespa.zookeeper.server.VespaZooKeeperServer;
 import com.yahoo.yolean.Exceptions;
 
 import java.nio.file.Files;
